langchain/content-builder/pyproject.toml
[project]
name = "content-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.5.3",
    "langchain>=1.2.15",
    "langchain-ollama>=1.1.0",
    "ollama>=0.6.1",
    "pillow>=12.2.0",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
    "tavily-python>=0.7.23",
]
